backend/requirements.txt
# 🌐 Web framework & ASGI server
fastapi==0.116.1
starlette==0.47.3
uvicorn==0.35.0

# 🔗 HTTP clients & protocols
httpx==0.28.1
httpcore==1.0.9
h11==0.16.0
urllib3==2.0.0
python-multipart==0.0.20
multipart  # if used separately

# 🧠 AI & Azure SDK
openai==1.102.0
azure-core==1.35.0
azure-ai-documentintelligence  # version not shown

# 🧪 Typing, validation, and error handling
pydantic  # version not shown
pydantic-core==2.33.2
annotated-types==0.7.0
typing-extensions==4.15.0
typing-inspection==0.4.1
exceptiongroup==1.3.0
isodate==0.7.2
jiter==0.10.0

# 🌍 Encoding & compatibility
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
distro==1.9.0
six==1.17.0
sniffio==1.3.1

# 🔐 Environment config
python-dotenv==1.1.1

# 🧰 Packaging tools (optional for runtime)
pip==25.2
setuptools==80.9.0
wheel==0.45.1

# 🎯 Developer UX
tqdm==4.67.1
colorama==0.4.6
click==8.2.1